refactored code structure to follow prbench
