+++ src/prbench_bilevel_planning/env_models/geom2d/__init__.py
@@ -1,0 +1,5 @@
+"""Geom2D environment models for bilevel planning."""
+
+from .geom2d_utils import Geom2dRobotController
+
+__all__ = ["Geom2dRobotController"]
