+++ src/prbench_bilevel_planning/env_models/geom2d/obstruction2d.py
@@ -32,7 +32,9 @@ from relational_structs import (
 )
 from relational_structs.spaces import ObjectCentricBoxSpace, ObjectCentricStateSpace
 
-from prbench_bilevel_planning.env_models import Geom2dRobotController
+from prbench_bilevel_planning.env_models.geom2d.geom2d_utils import (
+    Geom2dRobotController,
+)
 
 
 def create_bilevel_planning_models(
