+++ tests/env_models/geom2d/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for geom2d environment models."""
