+++ tests/env_models/geom2d/test_stickbutton2d.py
